Breaking change: refactor app and add tests
add put/delete mappings, app is now throwing 400 or 404 error with custom template depending on error, add tests for most classes

+++ src/main/java/com/xmies/Library/controller/AdminLibraryController.java
@@ -7,11 +7,13 @@ import com.xmies.Library.entity.Review;
 import com.xmies.Library.service.LibraryService;
 import jakarta.validation.Valid;
 import org.springframework.beans.propertyeditors.StringTrimmerEditor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.WebDataBinder;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -54,17 +56,27 @@ public class AdminLibraryController {
         return "redirect:/library/list";
     }
 
-    @GetMapping("/updateBookForm")
+    @PutMapping("/updateBookForm")
     public String updateBook(@RequestParam("bookId") int id, Model model) {
-        Book book = libraryService.findBookAndAuthorsByBookId(id);
+
+        if (!libraryService.bookExistsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Unable to find book");
+        }
+
+        Book book = libraryService.findBookById(id);
 
         model.addAttribute("book", book);
 
         return "library/book-add-form";
     }
 
-    @GetMapping("/deleteBook")
+    @DeleteMapping("/deleteBook")
     public String deleteBook(@RequestParam("bookId") int id) {
+
+        if (!libraryService.bookExistsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Unable to find book");
+        }
+
         libraryService.deleteBookById(id);
 
         return "redirect:/library/list";
@@ -92,8 +104,13 @@ public class AdminLibraryController {
         return "redirect:/library/authorsList";
     }
 
-    @GetMapping("/updateAuthorForm")
+    @PutMapping("/updateAuthorForm")
     public String updateAuthor(@RequestParam("authorId") int id, Model model) {
+
+        if (!libraryService.authorExistsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Unable to find author");
+        }
+
         Author author = libraryService.findAuthorById(id);
 
         model.addAttribute("author", author);
@@ -101,8 +118,13 @@ public class AdminLibraryController {
         return "library/author-add-form";
     }
 
-    @GetMapping("/deleteAuthor")
+    @DeleteMapping("/deleteAuthor")
     public String deleteAuthor(@RequestParam("authorId") int id) {
+
+        if(!libraryService.authorExistsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Unable to find author");
+        }
+
         libraryService.deleteAuthorById(id);
 
         return "redirect:/library/authorsList";
@@ -110,8 +132,12 @@ public class AdminLibraryController {
 
     @GetMapping("/manageAuthors")
     public String listAuthors(@RequestParam("bookId") int id, Model model) {
-        List<Author> authors = libraryService.findAuthorsAndBookByBookId(id);
 
+        if (!libraryService.bookExistsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Unable to find book");
+        }
+
+        List<Author> authors = libraryService.findAuthorsAndBookByBookId(id);
         model.addAttribute("authors", authors);
 
         return "library/list-authors-for-book";
@@ -119,6 +145,10 @@ public class AdminLibraryController {
 
     @GetMapping("bindAuthorToBookList")
     public String bindAuthorToBookList(@RequestParam("authorId") int id, Model model) {
+
+        if (!libraryService.authorExistsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Unable to find author");
+        }
         List<Book> books = libraryService.findAllBooks();
         Author author = libraryService.findAuthorById(id);
 
@@ -128,28 +158,40 @@ public class AdminLibraryController {
         return "library/bind-author-to-book";
     }
 
-    @GetMapping("bindAuthorToBook")
-    public String bindAuthorToBook(@RequestParam("authorId") int id, @RequestParam("bookId") int bookId) {
-        libraryService.bindAuthorToBook(id, bookId);
+    @PutMapping("bindAuthorToBook")
+    public String bindAuthorToBook(@RequestParam("authorId") int authorId, @RequestParam("bookId") int bookId) {
+
+        if (!libraryService.bookExistsById(bookId) || !libraryService.authorExistsById(authorId)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Unable to find given author and/or book");
+        }
+
+        libraryService.bindAuthorToBook(authorId, bookId);
 
         return "redirect:/library/menu";
     }
 
     @GetMapping("manageAuthorDetails")
     public String manageAuthorDetails(@RequestParam("authorId") int id, Model model) {
-        Author author = libraryService.findAuthorById(id);
 
+        if (!libraryService.authorExistsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Unable to find author");
+        }
+
+        Author author = libraryService.findAuthorById(id);
         model.addAttribute("author", author);
 
         return "library/author-details-manager";
     }
 
-    @GetMapping("updateAuthorDetails")
+    @PutMapping("updateAuthorDetails")
     public String updateAuthorDetails(@RequestParam("authorId") int id, Model model) {
-        Author author = libraryService.findAuthorById(id);
-        AuthorDetails authorDetails = author.getAuthorDetails();
 
-        model.addAttribute("authorDetails", authorDetails);
+        if (!libraryService.authorExistsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Unable to find author");
+        }
+
+        Author author = libraryService.findAuthorAndAuthorDetailById(id);
+        model.addAttribute("authorDetails", author.getAuthorDetails());
 
         return "library/author-details-add-form";
     }
@@ -163,25 +205,41 @@ public class AdminLibraryController {
             return "library/author-details-add-form";
         }
 
-        libraryService.save(authorDetails);
+        if (!libraryService.authorExistsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Unable to find author");
+        }
+
+        Author author = libraryService.findAuthorById(id);
+        author.setAuthorDetails(authorDetails);
+
+        libraryService.save(author);
 
         return "redirect:/library/seeAuthorDetails?authorId=" + id;
     }
 
-    @GetMapping("reviewUpdateForm")
+    @PutMapping("reviewUpdateForm")
     public String reviewUpdateForm(@RequestParam("bookId") int bookId,
-                                   @RequestParam("reviewId") int id,
+                                   @RequestParam("reviewId") int reviewId,
                                    Model model) {
 
-        Review review = libraryService.findReviewById(id);
+        if (!libraryService.bookExistsById(bookId) || !libraryService.reviewExistsById(reviewId)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Unable to find book and/or review");
+        }
+
+        Review review = libraryService.findReviewById(reviewId);
 
         model.addAttribute("review", review);
 
         return "library/review-add-form";
     }
 
-    @GetMapping("deleteReview")
+    @DeleteMapping("deleteReview")
     public String deleteReview(@RequestParam("bookId") int bookId, @RequestParam("reviewId") int id) {
+
+        if (!libraryService.reviewExistsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Unable to find review");
+        }
+
         libraryService.deleteReviewById(id);
 
         return "redirect:/library/book-information?bookId=" + bookId;
